% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{orthogonal_projection}
\alias{orthogonal_projection}
\title{Orthogonal Projection using Trigonometric Basis}
\usage{
orthogonal_projection(X, Y, M)
}
\arguments{
\item{X}{A numeric vector of predictor values.}

\item{Y}{A numeric vector of response values.}

\item{M}{An integer specifying the maximum order of the trigonometric basis functions.}
}
\value{
A linear model object.
}
\description{
Fits a linear model using trigonometric basis functions up to the specified order \code{M}.
}
\examples{
X <- seq(0, 1, length.out = 100)
Y <- 10 * (X - 0.5)^2 + rnorm(100)
M <- 5
fit <- orthogonal_projection(X, Y, M)
summary(fit)
}
