% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNN_estimation.R
\name{plot_knn_estimates}
\alias{plot_knn_estimates}
\title{Plot kNN Estimates}
\usage{
plot_knn_estimates(X, Y, X_grid, k_values, g_star)
}
\arguments{
\item{X}{Numeric vector of predictor values.}

\item{Y}{Numeric vector of response values.}

\item{X_grid}{Numeric vector of points for plotting the estimates.}

\item{k_values}{Numeric vector of k values to use for kNN estimation.}

\item{g_star}{Function representing the true function to compare against.}
}
\value{
A ggplot object showing the kNN estimates and the true function.
}
\description{
Plots the kNN estimates for given values of k and compares them with the true function.
}
\examples{
# Define the true function
g_star <- function(x) { 10 * (x - 0.5)^2 + 2 }

# Generate data
set.seed(123)
X <- seq(0, 1, length.out = 100)
Y <- g_star(X) + rnorm(100, 0, 0.05)
X_grid <- seq(0, 1, length.out = 1000)

# Plot kNN estimates
plot_knn_estimates(X, Y, X_grid, c(1, 3, 15, 100), g_star)
}
