% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{trig_basis}
\alias{trig_basis}
\title{Trigonometric Basis Functions}
\usage{
trig_basis(x, M)
}
\arguments{
\item{x}{A numeric vector.}

\item{M}{An integer specifying the maximum order of the trigonometric basis functions.}
}
\value{
A matrix where each column represents a trigonometric basis function.
}
\description{
Creates a trigonometric basis for the input vector \code{x} up to the specified order \code{M}.
}
\examples{
x <- seq(0, 1, length.out = 100)
M <- 5
basis <- trig_basis(x, M)
}
